按照阿里巴巴规范 修复代码中不规范的部分

+++ src/main/java/com/linda/framework/rpc/filter/SimpleRpcFilterChain.java
@@ -26,7 +26,7 @@ public class SimpleRpcFilterChain implements RpcFilterChain {
             index = 0;
         }
         rpcFilterIndex.set(index + 1);
-        return index;
+        return index.intValue();
     }
 
     @Override
